only mongo remains

+++ dashboard/model/model.go
@@ -3,5 +3,5 @@ package model
 // TimeSeries is a struct that holds the data for the history of Trump's popularity
 type TimeSeries struct {
 	Timestamp       string
-	TrumpPopularity float64
+	TrumpPopularity int
 }
